+++ affinda/models/_models_py3.py
@@ -1743,6 +1743,8 @@ class Invoice(msrest.serialization.Model):
 class InvoiceData(msrest.serialization.Model):
     """InvoiceData.
 
+    :ivar tables:
+    :vartype tables: list[~affinda.models.InvoiceDataTablesItem]
     :ivar invoice_date:
     :vartype invoice_date: ~affinda.models.DateAnnotation
     :ivar invoice_order_date:
@@ -1816,6 +1818,7 @@ class InvoiceData(msrest.serialization.Model):
     """
 
     _attribute_map = {
+        "tables": {"key": "tables", "type": "[InvoiceDataTablesItem]"},
         "invoice_date": {"key": "invoiceDate", "type": "DateAnnotation"},
         "invoice_order_date": {"key": "invoiceOrderDate", "type": "DateAnnotation"},
         "payment_date_due": {"key": "paymentDateDue", "type": "DateAnnotation"},
@@ -1898,6 +1901,7 @@ class InvoiceData(msrest.serialization.Model):
     def __init__(
         self,
         *,
+        tables: Optional[List["InvoiceDataTablesItem"]] = None,
         invoice_date: Optional["DateAnnotation"] = None,
         invoice_order_date: Optional["DateAnnotation"] = None,
         payment_date_due: Optional["DateAnnotation"] = None,
@@ -1936,6 +1940,8 @@ class InvoiceData(msrest.serialization.Model):
         **kwargs,
     ):
         """
+        :keyword tables:
+        :paramtype tables: list[~affinda.models.InvoiceDataTablesItem]
         :keyword invoice_date:
         :paramtype invoice_date: ~affinda.models.DateAnnotation
         :keyword invoice_order_date:
@@ -2008,6 +2014,7 @@ class InvoiceData(msrest.serialization.Model):
         :paramtype supplier_website: ~affinda.models.InvoiceDataSupplierWebsite
         """
         super(InvoiceData, self).__init__(**kwargs)
+        self.tables = tables
         self.invoice_date = invoice_date
         self.invoice_order_date = invoice_order_date
         self.payment_date_due = payment_date_due
@@ -5003,6 +5010,26 @@ class InvoiceDataSupplierWebsite(
         self.is_verified = is_verified
         self.classification = classification
         self.parsed = parsed
+
+
+class InvoiceDataTablesItem(msrest.serialization.Model):
+    """InvoiceDataTablesItem.
+
+    :ivar rows:
+    :vartype rows: list[~affinda.models.RowAnnotation]
+    """
+
+    _attribute_map = {
+        "rows": {"key": "rows", "type": "[RowAnnotation]"},
+    }
+
+    def __init__(self, *, rows: Optional[List["RowAnnotation"]] = None, **kwargs):
+        """
+        :keyword rows:
+        :paramtype rows: list[~affinda.models.RowAnnotation]
+        """
+        super(InvoiceDataTablesItem, self).__init__(**kwargs)
+        self.rows = rows
 
 
 class Meta(msrest.serialization.Model):
@@ -9185,6 +9212,108 @@ class ResumeSkillSourcesItem(msrest.serialization.Model):
         super(ResumeSkillSourcesItem, self).__init__(**kwargs)
         self.section = section
         self.position = position
+
+
+class RowAnnotation(msrest.serialization.Model):
+    """RowAnnotation.
+
+    :ivar code:
+    :vartype code: str
+    :ivar date:
+    :vartype date: str
+    :ivar description:
+    :vartype description: str
+    :ivar unit:
+    :vartype unit: str
+    :ivar unit_price:
+    :vartype unit_price: float
+    :ivar quantity:
+    :vartype quantity: float
+    :ivar discount:
+    :vartype discount: str
+    :ivar base_total:
+    :vartype base_total: float
+    :ivar tax_rate:
+    :vartype tax_rate: str
+    :ivar tax_total:
+    :vartype tax_total: float
+    :ivar total:
+    :vartype total: float
+    :ivar other:
+    :vartype other: str
+    """
+
+    _attribute_map = {
+        "code": {"key": "code", "type": "str"},
+        "date": {"key": "date", "type": "str"},
+        "description": {"key": "description", "type": "str"},
+        "unit": {"key": "unit", "type": "str"},
+        "unit_price": {"key": "unitPrice", "type": "float"},
+        "quantity": {"key": "quantity", "type": "float"},
+        "discount": {"key": "discount", "type": "str"},
+        "base_total": {"key": "baseTotal", "type": "float"},
+        "tax_rate": {"key": "taxRate", "type": "str"},
+        "tax_total": {"key": "taxTotal", "type": "float"},
+        "total": {"key": "total", "type": "float"},
+        "other": {"key": "other", "type": "str"},
+    }
+
+    def __init__(
+        self,
+        *,
+        code: Optional[str] = None,
+        date: Optional[str] = None,
+        description: Optional[str] = None,
+        unit: Optional[str] = None,
+        unit_price: Optional[float] = None,
+        quantity: Optional[float] = None,
+        discount: Optional[str] = None,
+        base_total: Optional[float] = None,
+        tax_rate: Optional[str] = None,
+        tax_total: Optional[float] = None,
+        total: Optional[float] = None,
+        other: Optional[str] = None,
+        **kwargs,
+    ):
+        """
+        :keyword code:
+        :paramtype code: str
+        :keyword date:
+        :paramtype date: str
+        :keyword description:
+        :paramtype description: str
+        :keyword unit:
+        :paramtype unit: str
+        :keyword unit_price:
+        :paramtype unit_price: float
+        :keyword quantity:
+        :paramtype quantity: float
+        :keyword discount:
+        :paramtype discount: str
+        :keyword base_total:
+        :paramtype base_total: float
+        :keyword tax_rate:
+        :paramtype tax_rate: str
+        :keyword tax_total:
+        :paramtype tax_total: float
+        :keyword total:
+        :paramtype total: float
+        :keyword other:
+        :paramtype other: str
+        """
+        super(RowAnnotation, self).__init__(**kwargs)
+        self.code = code
+        self.date = date
+        self.description = description
+        self.unit = unit
+        self.unit_price = unit_price
+        self.quantity = quantity
+        self.discount = discount
+        self.base_total = base_total
+        self.tax_rate = tax_rate
+        self.tax_total = tax_total
+        self.total = total
+        self.other = other
 
 
 class SkillAnnotation(Annotation):